njit tests passing
Former-commit-id: 56c1912944c669cb5853dae47724fa8ec632c666 [formerly 004c8792aa7f3c04c3b95134866fbd9fbb9bff99]
Former-commit-id: ad67110577c4d93ceaa991fbe450dc8ede17d90c

+++ dyntapy/dta/core/route_choice/aon_setup.py
@@ -86,9 +86,9 @@ def incremental_loading(network: Network, time: SimulationTime, dynamic_demand: 
     for k in range(1,K+1):
         # update demand such that the current slice of demand is added.
         if k ==1 :
-            demand_factor = 1/K
+            demand_factor = np.float32(1/K)
         else:
-            demand_factor =  k/(k-1)
+            demand_factor =  np.float32(k/(k-1))
         for demand in dynamic_demand.demands:
             demand.to_destinations.values =  demand.to_destinations.values*demand_factor
             demand.to_origins.values = demand.to_origins.values * demand_factor
